+++ all/modules/azure_rm_sqlfirewallrule.py
@@ -142,6 +142,7 @@ class AzureRMFirewallRules(AzureRMModuleBase):
         for key in list(self.module_arg_spec.keys()) + ['tags']:
             if hasattr(self, key):
                 setattr(self, key, kwargs[key])
+            elif kwargs[key] is not None:
 
         old_response = None
         results = dict()
@@ -179,18 +180,21 @@ class AzureRMFirewallRules(AzureRMModuleBase):
                 self.results['changed'] = True
             else:
                 self.results['changed'] = old_response.__ne__(response)
-
-            # remove unnecessary fields from return state
-            self.results["id"] = response["id"]
             self.log("Creation / Update done")
         elif self.to_do == Actions.Delete:
             self.log("FirewallRules instance deleted")
+
+            if self.check_mode:
+                return self.results
+
             self.delete_firewallrules()
             self.results['changed'] = True
         else:
             self.log("FirewallRules instance unchanged")
-            self.results['state'] = old_response
             self.results['changed'] = False
+            response = old_response
+
+        self.results["id"] = response["id"]
 
         return self.results
