+++ all/modules/azure_rm_sqlservercommunicationlink.py
@@ -137,6 +137,7 @@ class AzureRMServerCommunicationLinks(AzureRMModuleBase):
         for key in list(self.module_arg_spec.keys()) + ['tags']:
             if hasattr(self, key):
                 setattr(self, key, kwargs[key])
+            elif kwargs[key] is not None:
 
         old_response = None
         results = dict()
@@ -174,19 +175,22 @@ class AzureRMServerCommunicationLinks(AzureRMModuleBase):
                 self.results['changed'] = True
             else:
                 self.results['changed'] = old_response.__ne__(response)
-
-            # remove unnecessary fields from return state
-            self.results["id"] = response["id"]
-            self.results["state"] = response["state"]
             self.log("Creation / Update done")
         elif self.to_do == Actions.Delete:
             self.log("ServerCommunicationLinks instance deleted")
+
+            if self.check_mode:
+                return self.results
+
             self.delete_servercommunicationlinks()
             self.results['changed'] = True
         else:
             self.log("ServerCommunicationLinks instance unchanged")
-            self.results['state'] = old_response
             self.results['changed'] = False
+            response = old_response
+
+        self.results["id"] = response["id"]
+        self.results["state"] = response["state"]
 
         return self.results
